fix: prevent public caching of profile-specific models
