+++ server/test/test_latest_mlp_model.py
@@ -1,6 +1,4 @@
-import os
-import socket
-import subprocess
+import time
 import time
 import urllib.request
 import urllib.error
@@ -13,62 +11,8 @@ import pytest
 SERVER_DIR = Path(__file__).resolve().parents[1]
 
 
-def _get_free_port() -> int:
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        s.bind(("127.0.0.1", 0))
-        return s.getsockname()[1]
-
-
-PORT = os.environ.get("PORT") or str(_get_free_port())
-HOST = "127.0.0.1"
-BASE_URL = f"http://{HOST}:{PORT}"
-
-def start_server():
-    env = os.environ.copy()
-    env.setdefault("API_TOKEN", "testtoken")
-    env.setdefault("PORT", PORT)
-    env.setdefault("HOST", HOST)
-    subprocess.run(
-        ["npm", "run", "build"],
-        cwd=SERVER_DIR,
-        env=env,
-        stdout=subprocess.DEVNULL,
-        stderr=subprocess.DEVNULL,
-        check=True,
-    )
-    proc = subprocess.Popen(
-        ["node", "dist/server.js"],
-        cwd=SERVER_DIR,
-        env=env,
-        stdout=subprocess.DEVNULL,
-        stderr=subprocess.DEVNULL,
-    )
-    start = time.time()
-    headers = {"Authorization": "Bearer testtoken"}
-    req = urllib.request.Request(f"{BASE_URL}/model-version", headers=headers)
-    while True:
-        if proc.poll() is not None:
-            raise RuntimeError("server failed to start")
-        try:
-            with urllib.request.urlopen(req, timeout=5) as resp:
-                if resp.getcode() == 200:
-                    break
-        except (urllib.error.URLError, ConnectionRefusedError, socket.timeout):
-
-            if time.time() - start > 30:
-                raise RuntimeError("server did not start in time")
-            time.sleep(0.5)
-    return proc
-
-def stop_server(proc):
-    proc.terminate()
-    try:
-        proc.wait(timeout=5)
-    except subprocess.TimeoutExpired:
-        proc.kill()
-
-def fetch_latest_mlp_model(profile_id=None, extra_headers=None):
-    url = f"{BASE_URL}/latest-mlp-model"
+def fetch_latest_mlp_model(base_url, profile_id=None, extra_headers=None):
+    url = f"{base_url}/latest-mlp-model"
     if profile_id:
         url += f"?profileId={profile_id}"
     headers = {"Authorization": "Bearer testtoken"}
@@ -80,14 +24,6 @@ def fetch_latest_mlp_model(profile_id=None, extra_headers=None):
             return resp.getcode()
     except urllib.error.HTTPError as e:
         return e.code
-
-@pytest.fixture
-def running_server():
-    proc = start_server()
-    try:
-        yield
-    finally:
-        stop_server(proc)
 
 @pytest.fixture
 def missing_data_dir():
@@ -124,23 +60,23 @@ def model_file():
         if backup_dir is not None and Path(backup_dir).exists():
             shutil.move(str(backup_dir), str(data_dir))
 
-def test_latest_mlp_model_requires_authorization(model_file, running_server):
-    status = fetch_latest_mlp_model(profile_id="p1")
+def test_latest_mlp_model_requires_authorization(model_file, running_server, base_url):
+    status = fetch_latest_mlp_model(base_url, profile_id="p1")
     assert status == 403
 
-def test_latest_mlp_model_returns_404_when_missing(missing_data_dir, running_server):
-    status = fetch_latest_mlp_model()
+def test_latest_mlp_model_returns_404_when_missing(missing_data_dir, running_server, base_url):
+    status = fetch_latest_mlp_model(base_url)
     assert status == 404
 
-def test_latest_mlp_model_returns_200_for_authorized_owner(model_file, running_server):
+def test_latest_mlp_model_returns_200_for_authorized_owner(model_file, running_server, base_url):
     status = fetch_latest_mlp_model(
-        profile_id="p1", extra_headers={"x-profile-id": "p1"}
+        base_url, profile_id="p1", extra_headers={"x-profile-id": "p1"}
     )
     assert status == 200
 
 
-def test_latest_mlp_model_sets_headers(model_file, running_server):
-    url = f"{BASE_URL}/latest-mlp-model?profileId=p1"
+def test_latest_mlp_model_sets_headers(model_file, running_server, base_url):
+    url = f"{base_url}/latest-mlp-model?profileId=p1"
     headers = {"Authorization": "Bearer testtoken", "x-profile-id": "p1"}
     req = urllib.request.Request(url, headers=headers)
     with urllib.request.urlopen(req, timeout=5) as resp:
@@ -153,6 +89,5 @@ def test_latest_mlp_model_sets_headers(model_file, running_server):
         version = resp.headers.get("X-Model-Version")
         assert version is not None and version.isdigit()
         cache_control = resp.headers.get("Cache-Control")
-        assert cache_control == "public, max-age=0, must-revalidate"
-        cdn_cache = resp.headers.get("CDN-Cache-Control")
-        assert cdn_cache == "max-age=3600"
+        assert cache_control == "private, max-age=0, must-revalidate"
+        assert "CDN-Cache-Control" not in resp.headers
